+++ generator/lib/builders/protocols/rest_json_builder.dart
@@ -1,4 +1,4 @@
-import 'package:aws_client.generator/builders/builder.dart';
+import 'package:aws_client.generator/builders/protocols/service_builder.dart';
 import 'package:aws_client.generator/model/api.dart';
 import 'package:aws_client.generator/model/operation.dart';
